# Blazor/OpenAPIs/swagger1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API Academica",
    "version": "v1"
  },
  "paths": {
    "/AlumnosInscripciones": {
      "get": {
        "tags": [
          "AlumnosInscripciones"
        ],
        "operationId": "Get All Alumnos Inscripciones",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlumnosInscripcione"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlumnosInscripcione"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlumnosInscripcione"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AlumnosInscripciones"
        ],
        "operationId": "Add Alumno Inscripcion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlumnosInscripcione"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AlumnosInscripcione"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AlumnosInscripcione"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "AlumnosInscripciones"
        ],
        "operationId": "Update Alumno Inscripcion",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlumnosInscripcione"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AlumnosInscripcione"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AlumnosInscripcione"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/AlumnosInscripciones/curso/{idCurso}": {
      "get": {
        "tags": [
          "AlumnosInscripciones"
        ],
        "operationId": "Get Alumnos Inscripciones By Curso",
        "parameters": [
          {
            "name": "idCurso",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlumnosInscripcione"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlumnosInscripcione"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlumnosInscripcione"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/AlumnosInscripciones/alumno/{idAlumno}": {
      "get": {
        "tags": [
          "AlumnosInscripciones"
        ],
        "operationId": "Get Alumnos Inscripciones By Alumno",
        "parameters": [
          {
            "name": "idAlumno",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlumnosInscripcione"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlumnosInscripcione"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlumnosInscripcione"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/AlumnosInscripciones/{id}": {
      "get": {
        "tags": [
          "AlumnosInscripciones"
        ],
        "operationId": "Get Alumno Inscripcion",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "AlumnosInscripciones"
        ],
        "operationId": "Delete Alumno Inscripcion",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlumnosInscripcione": {
        "type": "object",
        "properties": {
          "idInscripcion": {
            "type": "integer",
            "format": "int32"
          },
          "idAlumno": {
            "type": "integer",
            "format": "int32"
          },
          "idCurso": {
            "type": "integer",
            "format": "int32"
          },
          "condicion": {
            "type": "string",
            "nullable": true
          },
          "nota": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idAlumnoNavigation": {
            "$ref": "#/components/schemas/Persona"
          },
          "idCursoNavigation": {
            "$ref": "#/components/schemas/Curso"
          }
        },
        "additionalProperties": false
      },
      "Comisione": {
        "type": "object",
        "properties": {
          "idComision": {
            "type": "integer",
            "format": "int32"
          },
          "descComision": {
            "type": "string",
            "nullable": true
          },
          "anioEspecialidad": {
            "type": "integer",
            "format": "int32"
          },
          "idPlan": {
            "type": "integer",
            "format": "int32"
          },
          "cursos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Curso"
            },
            "nullable": true
          },
          "idPlanNavigation": {
            "$ref": "#/components/schemas/Plane"
          }
        },
        "additionalProperties": false
      },
      "Curso": {
        "type": "object",
        "properties": {
          "idCurso": {
            "type": "integer",
            "format": "int32"
          },
          "idMateria": {
            "type": "integer",
            "format": "int32"
          },
          "idComision": {
            "type": "integer",
            "format": "int32"
          },
          "anioCalendario": {
            "type": "integer",
            "format": "int32"
          },
          "cupo": {
            "type": "integer",
            "format": "int32"
          },
          "alumnosInscripciones": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlumnosInscripcione"
            },
            "nullable": true
          },
          "docentesCursos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocentesCurso"
            },
            "nullable": true
          },
          "idComisionNavigation": {
            "$ref": "#/components/schemas/Comisione"
          },
          "idMateriaNavigation": {
            "$ref": "#/components/schemas/Materia"
          }
        },
        "additionalProperties": false
      },
      "DocentesCurso": {
        "type": "object",
        "properties": {
          "idDictado": {
            "type": "integer",
            "format": "int32"
          },
          "idCurso": {
            "type": "integer",
            "format": "int32"
          },
          "idDocente": {
            "type": "integer",
            "format": "int32"
          },
          "cargo": {
            "type": "string",
            "nullable": true
          },
          "idCursoNavigation": {
            "$ref": "#/components/schemas/Curso"
          },
          "idDocenteNavigation": {
            "$ref": "#/components/schemas/Persona"
          }
        },
        "additionalProperties": false
      },
      "Especialidade": {
        "type": "object",
        "properties": {
          "idEspecialidad": {
            "type": "integer",
            "format": "int32"
          },
          "descEspecialidad": {
            "type": "string",
            "nullable": true
          },
          "planes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Plane"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Materia": {
        "type": "object",
        "properties": {
          "idMateria": {
            "type": "integer",
            "format": "int32"
          },
          "descMateria": {
            "type": "string",
            "nullable": true
          },
          "hsSemanales": {
            "type": "integer",
            "format": "int32"
          },
          "hsTotales": {
            "type": "integer",
            "format": "int32"
          },
          "idPlan": {
            "type": "integer",
            "format": "int32"
          },
          "cursos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Curso"
            },
            "nullable": true
          },
          "idPlanNavigation": {
            "$ref": "#/components/schemas/Plane"
          },
          "cantidadAlumnos": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Persona": {
        "type": "object",
        "properties": {
          "idPersona": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "apellido": {
            "type": "string",
            "nullable": true
          },
          "direccion": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "telefono": {
            "type": "string",
            "nullable": true
          },
          "fechaNac": {
            "type": "string",
            "format": "date-time"
          },
          "legajo": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "tipoPersona": {
            "type": "string",
            "nullable": true
          },
          "idPlan": {
            "type": "integer",
            "format": "int32"
          },
          "nombreUsuario": {
            "type": "string",
            "nullable": true
          },
          "clave": {
            "type": "string",
            "nullable": true
          },
          "alumnosInscripciones": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlumnosInscripcione"
            },
            "nullable": true
          },
          "docentesCursos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocentesCurso"
            },
            "nullable": true
          },
          "idPlanNavigation": {
            "$ref": "#/components/schemas/Plane"
          }
        },
        "additionalProperties": false
      },
      "Plane": {
        "type": "object",
        "properties": {
          "idPlan": {
            "type": "integer",
            "format": "int32"
          },
          "descPlan": {
            "type": "string",
            "nullable": true
          },
          "idEspecialidad": {
            "type": "integer",
            "format": "int32"
          },
          "comisiones": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comisione"
            },
            "nullable": true
          },
          "idEspecialidadNavigation": {
            "$ref": "#/components/schemas/Especialidade"
          },
          "materia": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Materia"
            },
            "nullable": true
          },
          "personas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Persona"
            },
            "nullable": true
          },
          "cantidadAlumnos": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Por favor ingrese el token JWT con el prefijo 'Bearer '",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}